requirements.txt
flask
flask_login
datetime
mysql-connector-python
configparser
pathlib
numpy
pandas
pymysql